+++ src/components/hexViewer/ConnectionsListItem.vue
@@ -2,6 +2,7 @@
   <q-item  @click.native="(event) => { itemClickHandler(index, item, event) }">
     <q-item-main>
       <q-item-tile label>{{peer}}</q-item-tile>
+      <q-item-tile style="font-size: 0.7rem" sublabel>{{ident}}</q-item-tile>
       <q-item-tile sublabel>{{date.formatDate(timestamp * 1000, 'DD/MM/YYYY HH:mm:ss')}}</q-item-tile>
     </q-item-main>
     <q-item-side right>
@@ -24,6 +25,7 @@ export default {
     return {
       date: date,
       timestamp: this.item.messages[0].timestamp,
+      ident: this.item.messages[0].ident,
       peer: this.item.peer
     }
   },